requirements.txt
absl-py==1.3.0
amqp==5.1.1
astunparse==1.6.3
async-timeout==4.0.2
billiard==3.6.4.0
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flatbuffers==22.12.6
flower==1.2.0
fonttools==4.38.0
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.8
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
humanize==4.4.0
idna==3.4
importlib-metadata==5.2.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
kombu==5.2.4
libclang==14.0.6
locust==2.14.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
msgpack==1.0.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.5
Pillow==9.3.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
rsa==4.9
scikit-learn==0.22
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.45
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
